db: move snapshots and wal directories to trash (#426)
* wal: fix metrics typo

* db: move snapshots and wal directories to trash

Previously, db.Close would call os.RemoveAll on the snapshots and wal
directories. This is an O(n) operation, which could lead to some leftover
files if we exceed the graceful shutdown period.

This commit changes the removal to move both directories to a trash/ directory,
which is an O(1) operation. This trash/ directory can be removed safely on
startup if it still exists.

Additionally, this commit moves closeInternal to before removing the
directories. This ensures that the WAL is flushed and closed before deletion.

+++ db.go
@@ -376,6 +376,9 @@ func (s *ColumnStore) DB(ctx context.Context, name string) (*DB, error) {
 	}
 
 	if dbSetupErr := func() error {
+		if err := os.RemoveAll(db.trashDir()); err != nil {
+			return err
+		}
 		db.txPool = NewTxPool(db.highWatermark)
 		db.compactorPool = newCompactorPool(db, s.compactionConfig)
 		if len(db.sources) != 0 {
@@ -465,12 +468,21 @@ func (db *DB) openWAL(ctx context.Context) (WAL, error) {
 	return wal, nil
 }
 
+const (
+	walPath       = "wal"
+	snapshotsPath = "snapshots"
+)
+
 func (db *DB) walDir() string {
-	return filepath.Join(db.storagePath, "wal")
+	return filepath.Join(db.storagePath, walPath)
 }
 
 func (db *DB) snapshotsDir() string {
-	return filepath.Join(db.storagePath, "snapshots")
+	return filepath.Join(db.storagePath, snapshotsPath)
+}
+
+func (db *DB) trashDir() string {
+	return filepath.Join(db.storagePath, "trash")
 }
 
 // recover attempts to recover database state from a combination of snapshots
@@ -767,17 +779,49 @@ func (db *DB) Close() error {
 		}
 	}
 
+	if err := db.closeInternal(); err != nil {
+		return err
+	}
+
 	if len(db.sinks) != 0 {
 		// If we've successfully persisted all the table blocks we can remove
-		// the wal and snapshots.
-		if err := os.RemoveAll(db.snapshotsDir()); err != nil {
-			return err
+		// the wal and snapshots. Move the two directories to a trash dir since
+		// it is an O(1) operation. The trash dir is cleaned up on startup in
+		// case we cannot delete the WAL/snapshots in time.
+		trashDir := db.trashDir()
+
+		if moveErr := func() error {
+			_ = os.Mkdir(trashDir, os.FileMode(0o755))
+			// Create a temporary directory in the trash dir to avoid clashing
+			// with other wal/snapshot dirs that might not have been removed
+			// previously.
+			tmpPath, err := os.MkdirTemp(trashDir, "")
+			if err != nil {
+				return err
+			}
+			if err := os.Rename(db.snapshotsDir(), filepath.Join(tmpPath, snapshotsPath)); err != nil && !os.IsNotExist(err) {
+				return err
+			}
+			if err := os.Rename(db.walDir(), filepath.Join(tmpPath, walPath)); err != nil && !os.IsNotExist(err) {
+				return err
+			}
+			return nil
+		}(); moveErr != nil {
+			// If we failed to move the wal/snapshots to the trash dir, fall
+			// back to attempting to remove them with RemoveAll.
+			if err := os.RemoveAll(db.snapshotsDir()); err != nil {
+				return fmt.Errorf("%v: %v", moveErr, err)
+			}
+			if err := os.RemoveAll(db.walDir()); err != nil {
+				return fmt.Errorf("%v: %v", moveErr, err)
+			}
+			return moveErr
 		}
-		if err := os.RemoveAll(db.walDir()); err != nil {
+		if err := os.RemoveAll(trashDir); err != nil {
 			return err
 		}
 	}
-	return db.closeInternal()
+	return nil
 }
 
 func (db *DB) closeInternal() error {
